==> 3_BhavCopy_Download_yfin.ipynb <==
# At first I though I will download all the data from NSE site.. 
# but there are better APIs which will reduce the workload by a ton. Moving to that.

import pandas as pd
import numpy as np
import requests 
#from pynse import *
#nse=Nse()
from nsepy.history import get_price_list
import time
import yfinance as yf

#Earlier the numbers were coming from bhavcopy, but those numbers are not adjusted
# Shifting to yfinance API

# Created a function to get the list of all dates
def update_date(date_fetch,increment):
    date_fetch = date_fetch + datetime.timedelta(days=increment)
    return date_fetch

#List of companies
df_companies = pd.read_csv('MCAP31032021_0.csv').set_index('SrNo')
df_companies = df_companies[df_companies['MarketCap'] > 0]
selected_tikrs = df_companies['Symbol'].to_list()
selected_tikrs = [s + '.NS' for s in selected_tikrs]
selected_tikrs.append('^NSEI')
selected_tikrs

import datetime
from datetime import date
import calendar
#prices = get_price_list(dt=date(2012,2,1))
start_date = '2010-01-01'
end_date = '2010-01-31'
#price = yf.download('HDFCLIFE.NS',start = '2019-1-1', end = '2021-1-1')['Adj Close']
price = yf.download(selected_tikrs[:5],start = start_date, end = end_date,interval = "1wk")[['Adj Close','Volume']] #


#pd.concat([price,price]).to_csv('testing_bhav.csv')

#nse.bhavcopy(dt.date(2020,1,1)) # this did not work out as the historical data was not there
import datetime
from datetime import date
import calendar
#prices = get_price_list(dt=date(2015,1,1))

# var_year = 2021
# var_month = 2
# mid_date  =  date(var_year, var_month, 15)
# start_date = mid_date.replace(day = 1)
# print(start_date)
#df_start = get_price_list(dt=date(var_year,var_month,int("{:%d}".format(start_date))))
#df_start =  get_price_list(dt=date(2021,2,3))
#df_start.head(4)

# Function to download bhavcopy of 1st date of each month
def getting_yfinance_data():
    #print(var_year,var_month)
    start_date = '2010-01-01'
    end_date = '2021-12-31'
    df_start = yf.download(selected_tikrs,start = start_date, end = end_date,interval = "1wk")[['Adj Close','Volume']] #
    print('Completed - ',start_date)
    df_start.columns = ['_'.join(col) for col in df_start.columns]
   #df_start.to_csv('df_bhavcopy_backup.csv')
    return df_start

df_bhavcopy = pd.DataFrame()
df_bhavcopy = getting_yfinance_data()
# Removing duplicates to remove same set of rows for same dates
df_bhavcopy = df_bhavcopy.drop_duplicates()
df_bhavcopy.to_csv('df_bhavcopy_yfin.csv')

df_bhavcopy

#getting_dates()
# print('today date',str(all_dates[0][0]))
# print("We need this date in this format {:%d, %b %Y}".format(all_dates[0][0]))
# print("We need this date in this format {:%d-%m-%Y}".format(all_dates[0][0]))
# print("Year is - {:%Y}".format(all_dates[0][0]))
# print("Month is - {:%b}".format(all_dates[0][0]))
# print("filenameis - {:%d%b%Y}".format(all_dates[0][0]).upper())


df_bhavcopy_yfin = pd.read_csv('df_bhavcopy_yfin.csv')
df_bhavcopy_melt = pd.melt(df_bhavcopy_yfin, id_vars=["Date"])
df_bhavcopy_melt = df_bhavcopy_melt[df_bhavcopy_melt['value'].notna()]
df_bhavcopy_melt[['KPI','Tikr']] = df_bhavcopy_melt['variable'].str.split('_',1, expand=True)
df_bhavcopy_melt = df_bhavcopy_melt.drop(['variable'],axis=1)
df_bhavcopy_melt.columns = ['TIMESTAMP','CLOSE','KPI','SYMBOL']

df_bhavcopy_melt

# #df_nifty.shape
# df_nifty = data.reset_index()
# df_nifty.columns = ['TIMESTAMP','OPEN','HIGH','LOW','CLOSE','ADJ_CLOSE','TOTTRDQTY']
# df_nifty.head()
df_bhavcopy_melt['TIMESTAMP'] = pd.DatetimeIndex(df_bhavcopy_melt['TIMESTAMP']).strftime('%d-%b-%Y').str.upper()
df_bhavcopy_melt.head()

df_bhav_pivot = pd.pivot_table(df_bhavcopy_melt,index=['TIMESTAMP','SYMBOL'], columns=['KPI'], values=['CLOSE'], aggfunc='sum')
df_bhav_pivot = df_bhav_pivot.reset_index()
df_bhav_pivot.columns = ['_'.join(col) for col in df_bhav_pivot.columns]
df_bhav_pivot.columns = ['TIMESTAMP','SYMBOL','CLOSE','Volume']
df_bhav_pivot

df_bhav_pivot.to_csv('df_bhavcopy.csv')
