==> 4_Bhavcopy_Processing.ipynb <==
import pandas as pd
import numpy as np
import re

df_bhavcopy = pd.read_csv('df_bhavcopy.csv').set_index('Unnamed: 0')
df_bhavcopy.head(4)

# df_bhavcopy['year'] = pd.DatetimeIndex(df_bhavcopy['TIMESTAMP']).year
# df_bhavcopy['month'] = pd.DatetimeIndex(df_bhavcopy['TIMESTAMP']).month
# df_bhavcopy['Yr_Month'] = pd.DatetimeIndex(df_bhavcopy['TIMESTAMP'],format="%d/%b/%Y %H:%M:%S").strftime('%Y_%m')
df_bhavcopy['Yr_Month'] = pd.to_datetime(df_bhavcopy['TIMESTAMP'],format="%d-%b-%Y")
df_bhavcopy['Yr_Month'] = df_bhavcopy['Yr_Month'].apply(lambda x: x.strftime('%Y_%m'))
df_bhavcopy['DATE'] = pd.to_datetime(df_bhavcopy['TIMESTAMP'])
df_bhavcopy.head()

df_bhavcopy.head(2)

# filepath_ca = 'C:\\Saiyad ADT\\Learnnig\\Sensex\\Screener1\\Data_2021_1121\\Corporate_Actions_BSE_2017_2020.csv'    
# df_split = pd.read_csv(filepath_ca, header=0)
# df_split.head(3)

df_bhavcopy[df_bhavcopy['SYMBOL'] == '^NSEI']

#Getting average of the 4 values
# Earlier we used to have split adjustments separately, since we shifted to yfinance so we dont that anymore
df_bhav_split_adjusted = df_bhavcopy.copy()
#df_bhav_split_adjusted['AVG'] = df_bhav_split_adjusted[['CLOSE']].mean(axis=1)

df_bhav_pivot = pd.pivot_table(df_bhav_split_adjusted, index=['SYMBOL'], columns=['Yr_Month'], values=['CLOSE'], aggfunc='mean')
# Replacing spaces from the column names and joining the two layer of column names into one.
df_bhav_pivot.columns =[str(s2) + str('_') + s1  for (s1,s2) in df_bhav_pivot.columns.tolist()]
df_bhav_pivot.reset_index(inplace=True)
# Reordering columns by name so that all similar column get together
df_bhav_pivot = df_bhav_pivot.reindex(sorted(df_bhav_pivot.columns), axis=1)
df_bhav_pivot.head(3)

df_bhav_yearly = pd.DataFrame()
for i in range(2009,2022):
    #print(i)
    df_bhav_yearly[i] = df_bhav_pivot.filter(regex=str(i)).mean(axis=1)
df_bhav_yearly['Tikr'] = df_bhav_pivot['SYMBOL']
df_bhav_yearly.head(10).to_clipboard()
df_bhav_yearly.head()

df_bhav_yearly.head().to_clipboard()

#df_bhav_yearly[df_bhav_yearly['Tikr'] == '^NSEI']
df_bhav_yearly.head()

len(df_bhav_yearly.columns) - 2
cols = df_bhav_yearly.columns
# Percentae calclation
def percentage_change(col1,col2):
    try:
        out = ((col2 - col1) / col1) * 100
    except:
        out = np.nan
    return out
    #return ((df_sales.loc[:,col2] - df_sales.loc[:,col1])/df_sales.loc[:,col1]) * 100

df_bhav_delta = pd.DataFrame()

# Looping over columns to get the percentage change YoY
# Earlier we were using 3 yrs delta for model
# Shifting to 1 yr return now
for i in range(len(df_bhav_yearly.columns) - 2):
    #print(cols[i])
    if i >= 2:
        df_bhav_delta[cols[i]] = percentage_change(df_bhav_yearly[cols[i]],df_bhav_yearly[cols[i+2]])

df_bhav_delta['Tikr'] = df_bhav_yearly['Tikr'].str.replace('^NSEI','Nifty50')
df_bhav_delta.iloc[-1,len(df_bhav_delta.columns)-1] = 'Nifty50'
#df_bhav_delta['Tikr'] = df_bhav_delta['Tikr'].str.replace('.NS','')
df_bhav_delta['Tikr'] = df_bhav_delta['Tikr'].str[:-3]
df_bhav_delta.head(4)
#df_bhav_delta.tail(4)

# # Averaging the years month over month changes to get an average return

# df_bhav_delta_monthly = pd.DataFrame()
# for i in range(2015,2022):
#     #print(i)
#     df_bhav_delta_monthly[i] = df_bhav_delta.filter(regex=str(i)).mean(axis=1)

# #df_bhav_delta[''] = df.filter(regex='spike')
# df_bhav_delta_monthly['Tikr'] = df_bhav_delta['Tikr']
# df_bhav_delta_monthly.head(4)
df_bhav_delta.to_csv('df_bhav_2yr_delta.csv')

# Initially we were calculating the MoM returns and averaging them. but that did not give any result in modeling..
# So now we moved to 3 year returns, expecting, this will reduce the noise in the system

# Normalization by subtracting the index growth
#df_bhav_delta[df_bhav_delta['Tikr'] == 'WIPRO']
df_bhav_delta[df_bhav_delta['Tikr'] == 'Nifty50']
# These numbers are pretty high, was the index really performing this well YoY ?

df_bhav_yearly[df_bhav_yearly['Tikr'] == 'WIPRO.NS']

df_bhav_delta[df_bhav_delta['Tikr'] == 'RELIANCE']    

#df_bhav_delta[df_bhav_delta['Tikr'] == 'TTKPRESTIG']  
df_bhav_delta[df_bhav_delta['Tikr'] == 'Nift']    

# Getting MoM return trend for each ticker
df_bhav_pivot.columns
df_bhav_monthly = pd.DataFrame()
# for i in range(2014,2022):
#     #print(i)
#     df_bhav_monthly[i] = df_bhav_pivot.filter(regex=str(i)).mean(axis=1)
# df_bhav_yearly['Tikr'] = df_bhav_pivot['SYMBOL']
# df_bhav_yearly.head(10).to_clipboard()
# df_bhav_yearly.head()

def percentage_change(col1,col2):
    try:
        out = ((col2 - col1) / col1) * 100
    except:
        out = np.nan
    return out
    
cols = df_bhav_pivot.columns
def monthly_return(df):
    for i in range(len(df_bhav_pivot.columns) - 1):
        #print(cols[i])
        if i >= 1:
            df_bhav_monthly[cols[i]] = percentage_change(df[cols[i-1]],df[cols[i]])
    return df_bhav_monthly

df_bhav_monthly_delta = monthly_return(df_bhav_pivot)
df_bhav_monthly_delta['Tikr'] = df_bhav_pivot['SYMBOL'].str.replace('^NSEI','Nifty50')
#df_bhav_monthly_delta['Tikr'] = df_bhav_monthly_delta['Tikr'].str.replace('^NSEI','Nifty50')
df_bhav_monthly_delta.iloc[-1,len(df_bhav_monthly_delta.columns)-1] = 'Nifty50'
df_bhav_monthly_delta['Tikr'] = df_bhav_monthly_delta['Tikr'].str[:-3]
df_bhav_monthly_delta.to_csv('df_bhav_monthly_delta.csv')


# Testing the data
df_bhav_pivot[df_bhav_pivot['SYMBOL'] == '20MICRONS.NS']

df_bhav_monthly_delta[df_bhav_monthly_delta['Tikr'] == '20MICRONS']

df_bhav_monthly_delta
