==> 5_Merging_Bhav_Fundamental_all.ipynb <==
import pandas as pd
import numpy as np

#df_qtrly = pd.read_csv('df_qtrly_sales_flat.csv').set_index('Unnamed: 0')
df_annual = pd.read_csv('df_annual_sales_flat.csv').set_index('Unnamed: 0')
df_balancesheet = pd.read_csv('df_other_fundamentals.csv').set_index('Unnamed: 0')

# This is not monthly but 3 yrs delta return
#df_bhav_return = pd.read_csv('df_bhav_1yr_delta.csv').set_index('Unnamed: 0')
df_bhav_return = pd.read_csv('df_bhav_2yr_delta.csv').set_index('Unnamed: 0')
df_bhav_monthly_delta = pd.read_csv('df_bhav_monthly_delta.csv').set_index('Unnamed: 0')

#df_qtrly.head(3)

df_annual.head(3)

df_bhav_monthly_delta.head(3)

df_bhav_return[df_bhav_return['Tikr'] == 'NIFTY50']

# Merging the three dataframes
#df_fundamentals = pd.merge(df_annual, df_qtrly, how='left',left_on=['Tikr'], right_on=['Tikr'])
df_fundamentals = pd.merge(df_annual, df_balancesheet, how='left',left_on=['Tikr'], right_on=['Tikr'])
df_merge =  pd.merge(df_fundamentals,df_bhav_return, how='left',left_on=['Tikr'], right_on=['Tikr'])
df_merge =  pd.merge(df_merge,df_bhav_monthly_delta, how='left',left_on=['Tikr'], right_on=['Tikr'])
pd.DataFrame(df_merge.columns).to_clipboard()
#pd.DataFrame(df_merge).to_clipboard()

# Identifying columns having null values
null_values = df_merge.isna().sum()/len(df_merge)
pd.DataFrame(null_values).to_clipboard()

df_bhav_return.columns

# Merging Industry

# As a next step we need to select the 5 year columns for training and the next 1 year return as y variable
# list_2013 = ['Tikr','Sales_2010','Sales_2011','Sales_2012',
#             'Sales_2011_YoY','Sales_2012_YoY',
#             'Revenue_2010','Revenue_2011','Revenue_2012',
#             'Revenue_2011_YoY','Revenue_2012_YoY',
#             'Expenses_2010','Expenses_2011','Expenses_2012',
#             'Expenses_2011_YoY','Expenses_2012_YoY',
#             'Net_Profit_2010','Net_Profit_2011','Net_Profit_2012',
#             'Net_Profit_2011_YoY','Net_Profit_2012_YoY',
#             'Operating_Profit_2010','Operating_Profit_2011','Operating_Profit_2012',
#             'Operating_Profit_2011_YoY','Operating_Profit_2012_YoY',
#             'EPS_in_Rs_2010','EPS_in_Rs_2011','EPS_in_Rs_2012',
#             'EPS_in_Rs_2011_YoY','EPS_in_Rs_2012_YoY',
#             'Tax__2010','Tax__2011','Tax__2012',
#             'Tax__2011_YoY','Tax__2012_YoY',
#             'OPM__2010','OPM__2011','OPM__2012',
#             'OPM__2011_YoY','OPM__2012_YoY','2015']

list_2015 = ['Tikr',
            'Sales_2010','Sales_2011','Sales_2012','Sales_2013','Sales_2014',
            'Sales_2011_YoY','Sales_2012_YoY','Sales_2013_YoY','Sales_2014_YoY',
            'Revenue_2010','Revenue_2011','Revenue_2012','Revenue_2013','Revenue_2014',
            'Revenue_2011_YoY','Revenue_2012_YoY','Revenue_2013_YoY','Revenue_2014_YoY',
            'Expenses_2010','Expenses_2011','Expenses_2012','Expenses_2013','Expenses_2014',
            'Expenses_2011_YoY','Expenses_2012_YoY','Expenses_2013_YoY','Expenses_2014_YoY',
            'Net_Profit_2010','Net_Profit_2011','Net_Profit_2012','Net_Profit_2013','Net_Profit_2014',
            'Net_Profit_2011_YoY','Net_Profit_2012_YoY','Net_Profit_2013_YoY','Net_Profit_2014_YoY',
            'Operating_Profit_2010','Operating_Profit_2011','Operating_Profit_2012','Operating_Profit_2013','Operating_Profit_2014',
            'Operating_Profit_2011_YoY','Operating_Profit_2012_YoY','Operating_Profit_2013_YoY','Operating_Profit_2014_YoY',
            'EPS_in_Rs_2010','EPS_in_Rs_2011','EPS_in_Rs_2012','EPS_in_Rs_2013','EPS_in_Rs_2014',
            'EPS_in_Rs_2011_YoY','EPS_in_Rs_2012_YoY','EPS_in_Rs_2013_YoY','EPS_in_Rs_2014_YoY',
            'Tax__2010','Tax__2011','Tax__2012','Tax__2013','Tax__2014',
            'Tax__2011_YoY','Tax__2012_YoY','Tax__2013_YoY','Tax__2014_YoY',
            'OPM__2010','OPM__2011','OPM__2012','OPM__2013','OPM__2014',
            'OPM__2011_YoY','OPM__2012_YoY','OPM__2013_YoY','OPM__2014_YoY',
            'Other_Assets_2010','Other_Assets_2011','Other_Assets_2012','Other_Assets_2013','Other_Assets_2014',
            'Other_Assets_2011_YoY','Other_Assets_2012_YoY','Other_Assets_2013_YoY','Other_Assets_2014_YoY',
            'Total_Liabilities_2010','Total_Liabilities_2011','Total_Liabilities_2012','Total_Liabilities_2013','Total_Liabilities_2014',
            'Total_Liabilities_2011_YoY','Total_Liabilities_2012_YoY','Total_Liabilities_2013_YoY','Total_Liabilities_2014_YoY',
            'Share_Capital_2010','Share_Capital_2011','Share_Capital_2012','Share_Capital_2013','Share_Capital_2014',
            'Share_Capital_2011_YoY','Share_Capital_2012_YoY','Share_Capital_2013_YoY','Share_Capital_2014_YoY',
            'Investments_2010','Investments_2011','Investments_2012','Investments_2013','Investments_2014',
            'Investments_2011_YoY','Investments_2012_YoY','Investments_2013_YoY','Investments_2014_YoY',
            'Borrowings_2010','Borrowings_2011','Borrowings_2012','Borrowings_2013','Borrowings_2014',
            'Borrowings_2011_YoY','Borrowings_2012_YoY','Borrowings_2013_YoY','Borrowings_2014_YoY',
            'Reserves_2010','Reserves_2011','Reserves_2012','Reserves_2013','Reserves_2014',
            'Reserves_2011_YoY','Reserves_2012_YoY','Reserves_2013_YoY','Reserves_2014_YoY',
            'Total_Assets_2010','Total_Assets_2011','Total_Assets_2012','Total_Assets_2013','Total_Assets_2014',
            'Total_Assets_2011_YoY','Total_Assets_2012_YoY','Total_Assets_2013_YoY','Total_Assets_2014_YoY',
            'Fixed_Assets_2010','Fixed_Assets_2011','Fixed_Assets_2012','Fixed_Assets_2013','Fixed_Assets_2014',
            'Fixed_Assets_2011_YoY','Fixed_Assets_2012_YoY','Fixed_Assets_2013_YoY','Fixed_Assets_2014_YoY',
            'Other_Liabilities_2010','Other_Liabilities_2011','Other_Liabilities_2012','Other_Liabilities_2013','Other_Liabilities_2014',
            'Other_Liabilities_2011_YoY','Other_Liabilities_2012_YoY','Other_Liabilities_2013_YoY','Other_Liabilities_2014_YoY',
            'CWIP_2010','CWIP_2011','CWIP_2012','CWIP_2013','CWIP_2014',
            'CWIP_2011_YoY','CWIP_2012_YoY','CWIP_2013_YoY','CWIP_2014_YoY',
            'Cash_from_Investing_Activity_2010','Cash_from_Investing_Activity_2011','Cash_from_Investing_Activity_2012','Cash_from_Investing_Activity_2013','Cash_from_Investing_Activity_2014',
            'Cash_from_Investing_Activity_2011_YoY','Cash_from_Investing_Activity_2012_YoY','Cash_from_Investing_Activity_2013_YoY','Cash_from_Investing_Activity_2014_YoY',
            'Cash_from_Operating_Activity_2010','Cash_from_Operating_Activity_2011','Cash_from_Operating_Activity_2012','Cash_from_Operating_Activity_2013','Cash_from_Operating_Activity_2014',
            'Cash_from_Operating_Activity_2011_YoY','Cash_from_Operating_Activity_2012_YoY','Cash_from_Operating_Activity_2013_YoY','Cash_from_Operating_Activity_2014_YoY',
            'Cash_from_Financing_Activity_2010','Cash_from_Financing_Activity_2011','Cash_from_Financing_Activity_2012','Cash_from_Financing_Activity_2013','Cash_from_Financing_Activity_2014',
            'Cash_from_Financing_Activity_2011_YoY','Cash_from_Financing_Activity_2012_YoY','Cash_from_Financing_Activity_2013_YoY','Cash_from_Financing_Activity_2014_YoY',
            'Net_Cash_Flow_2010','Net_Cash_Flow_2011','Net_Cash_Flow_2012','Net_Cash_Flow_2013','Net_Cash_Flow_2014',
            'Net_Cash_Flow_2011_YoY','Net_Cash_Flow_2012_YoY','Net_Cash_Flow_2013_YoY','Net_Cash_Flow_2014_YoY',
            'Share_Capital_%_2010','Share_Capital_%_2011','Share_Capital_%_2012','Share_Capital_%_2013','Share_Capital_%_2014',
            'Reserves_%_2010','Reserves_%_2011','Reserves_%_2012','Reserves_%_2013','Reserves_%_2014',
            'Borrowings_%_2010','Borrowings_%_2011','Borrowings_%_2012','Borrowings_%_2013','Borrowings_%_2014',
            'Other_Liabilities_%_2010','Other_Liabilities_%_2011','Other_Liabilities_%_2012','Other_Liabilities_%_2013','Other_Liabilities_%_2014',
            'Fixed_Assets_%_2010','Fixed_Assets_%_2011','Fixed_Assets_%_2012','Fixed_Assets_%_2013','Fixed_Assets_%_2014',
            'CWIP_%_2010','CWIP_%_2011','CWIP_%_2012','CWIP_%_2013','CWIP_%_2014',
            'Investments_%_2010','Investments_%_2011','Investments_%_2012','Investments_%_2013','Investments_%_2014',
            'Other_Assets_%_2010','Other_Assets_%_2011','Other_Assets_%_2012','Other_Assets_%_2013','Other_Assets_%_2014',
            'MarketCap','CapFlag','LageCap','MidCap','SmallCap','TinyCap','Industry',
            '2014_01_CLOSE','2014_02_CLOSE','2014_03_CLOSE','2014_04_CLOSE','2014_05_CLOSE','2014_06_CLOSE','2014_07_CLOSE','2014_08_CLOSE','2014_09_CLOSE','2014_10_CLOSE','2014_11_CLOSE','2014_12_CLOSE',
            '2015']

list_2016 = ['Tikr',
            'Sales_2011','Sales_2012','Sales_2013','Sales_2014','Sales_2015',
            'Sales_2012_YoY','Sales_2013_YoY','Sales_2014_YoY','Sales_2015_YoY',
            'Revenue_2011','Revenue_2012','Revenue_2013','Revenue_2014','Revenue_2015',
            'Revenue_2012_YoY','Revenue_2013_YoY','Revenue_2014_YoY','Revenue_2015_YoY',
            'Expenses_2011','Expenses_2012','Expenses_2013','Expenses_2014','Expenses_2015',
            'Expenses_2012_YoY','Expenses_2013_YoY','Expenses_2014_YoY','Expenses_2015_YoY',
            'Net_Profit_2011','Net_Profit_2012','Net_Profit_2013','Net_Profit_2014','Net_Profit_2015',
            'Net_Profit_2012_YoY','Net_Profit_2013_YoY','Net_Profit_2014_YoY','Net_Profit_2015_YoY',
            'Operating_Profit_2011','Operating_Profit_2012','Operating_Profit_2013','Operating_Profit_2014','Operating_Profit_2015',
            'Operating_Profit_2012_YoY','Operating_Profit_2013_YoY','Operating_Profit_2014_YoY','Operating_Profit_2015_YoY',
            'EPS_in_Rs_2011','EPS_in_Rs_2012','EPS_in_Rs_2013','EPS_in_Rs_2014','EPS_in_Rs_2015',
            'EPS_in_Rs_2012_YoY','EPS_in_Rs_2013_YoY','EPS_in_Rs_2014_YoY','EPS_in_Rs_2015_YoY',
            'Tax__2011','Tax__2012','Tax__2013','Tax__2014','Tax__2015',
            'Tax__2012_YoY','Tax__2013_YoY','Tax__2014_YoY','Tax__2015_YoY',
            'OPM__2011','OPM__2012','OPM__2013','OPM__2014','OPM__2015',
            'OPM__2012_YoY','OPM__2013_YoY','OPM__2014_YoY','OPM__2015_YoY',
            'Other_Assets_2011','Other_Assets_2012','Other_Assets_2013','Other_Assets_2014','Other_Assets_2015',
            'Other_Assets_2012_YoY','Other_Assets_2013_YoY','Other_Assets_2014_YoY','Other_Assets_2015_YoY',
            'Total_Liabilities_2011','Total_Liabilities_2012','Total_Liabilities_2013','Total_Liabilities_2014','Total_Liabilities_2015',
            'Total_Liabilities_2012_YoY','Total_Liabilities_2013_YoY','Total_Liabilities_2014_YoY','Total_Liabilities_2015_YoY',
            'Share_Capital_2011','Share_Capital_2012','Share_Capital_2013','Share_Capital_2014','Share_Capital_2015',
            'Share_Capital_2012_YoY','Share_Capital_2013_YoY','Share_Capital_2014_YoY','Share_Capital_2015_YoY',
            'Investments_2011','Investments_2012','Investments_2013','Investments_2014','Investments_2015',
            'Investments_2012_YoY','Investments_2013_YoY','Investments_2014_YoY','Investments_2015_YoY',
            'Borrowings_2011','Borrowings_2012','Borrowings_2013','Borrowings_2014','Borrowings_2015',
            'Borrowings_2012_YoY','Borrowings_2013_YoY','Borrowings_2014_YoY','Borrowings_2015_YoY',
            'Reserves_2011','Reserves_2012','Reserves_2013','Reserves_2014','Reserves_2015',
            'Reserves_2012_YoY','Reserves_2013_YoY','Reserves_2014_YoY','Reserves_2015_YoY',
            'Total_Assets_2011','Total_Assets_2012','Total_Assets_2013','Total_Assets_2014','Total_Assets_2015',
            'Total_Assets_2012_YoY','Total_Assets_2013_YoY','Total_Assets_2014_YoY','Total_Assets_2015_YoY',
            'Fixed_Assets_2011','Fixed_Assets_2012','Fixed_Assets_2013','Fixed_Assets_2014','Fixed_Assets_2015',
            'Fixed_Assets_2012_YoY','Fixed_Assets_2013_YoY','Fixed_Assets_2014_YoY','Fixed_Assets_2015_YoY',
            'Other_Liabilities_2011','Other_Liabilities_2012','Other_Liabilities_2013','Other_Liabilities_2014','Other_Liabilities_2015',
            'Other_Liabilities_2012_YoY','Other_Liabilities_2013_YoY','Other_Liabilities_2014_YoY','Other_Liabilities_2015_YoY',
            'CWIP_2011','CWIP_2012','CWIP_2013','CWIP_2014','CWIP_2015',
            'CWIP_2012_YoY','CWIP_2013_YoY','CWIP_2014_YoY','CWIP_2015_YoY',
            'Cash_from_Investing_Activity_2011','Cash_from_Investing_Activity_2012','Cash_from_Investing_Activity_2013','Cash_from_Investing_Activity_2014','Cash_from_Investing_Activity_2015',
            'Cash_from_Investing_Activity_2012_YoY','Cash_from_Investing_Activity_2013_YoY','Cash_from_Investing_Activity_2014_YoY','Cash_from_Investing_Activity_2015_YoY',
            'Cash_from_Operating_Activity_2011','Cash_from_Operating_Activity_2012','Cash_from_Operating_Activity_2013','Cash_from_Operating_Activity_2014','Cash_from_Operating_Activity_2015',
            'Cash_from_Operating_Activity_2012_YoY','Cash_from_Operating_Activity_2013_YoY','Cash_from_Operating_Activity_2014_YoY','Cash_from_Operating_Activity_2015_YoY',
            'Cash_from_Financing_Activity_2011','Cash_from_Financing_Activity_2012','Cash_from_Financing_Activity_2013','Cash_from_Financing_Activity_2014','Cash_from_Financing_Activity_2015',
            'Cash_from_Financing_Activity_2012_YoY','Cash_from_Financing_Activity_2013_YoY','Cash_from_Financing_Activity_2014_YoY','Cash_from_Financing_Activity_2015_YoY',
            'Net_Cash_Flow_2011','Net_Cash_Flow_2012','Net_Cash_Flow_2013','Net_Cash_Flow_2014','Net_Cash_Flow_2015',
            'Net_Cash_Flow_2012_YoY','Net_Cash_Flow_2013_YoY','Net_Cash_Flow_2014_YoY','Net_Cash_Flow_2015_YoY',
            'Share_Capital_%_2011','Share_Capital_%_2012','Share_Capital_%_2013','Share_Capital_%_2014','Share_Capital_%_2015',
            'Reserves_%_2011','Reserves_%_2012','Reserves_%_2013','Reserves_%_2014','Reserves_%_2015',
            'Borrowings_%_2011','Borrowings_%_2012','Borrowings_%_2013','Borrowings_%_2014','Borrowings_%_2015',
            'Other_Liabilities_%_2011','Other_Liabilities_%_2012','Other_Liabilities_%_2013','Other_Liabilities_%_2014','Other_Liabilities_%_2015',
            'Fixed_Assets_%_2011','Fixed_Assets_%_2012','Fixed_Assets_%_2013','Fixed_Assets_%_2014','Fixed_Assets_%_2015',
            'CWIP_%_2011','CWIP_%_2012','CWIP_%_2013','CWIP_%_2014','CWIP_%_2015',
            'Investments_%_2011','Investments_%_2012','Investments_%_2013','Investments_%_2014','Investments_%_2015',
            'Other_Assets_%_2011','Other_Assets_%_2012','Other_Assets_%_2013','Other_Assets_%_2014','Other_Assets_%_2015',
            'MarketCap','CapFlag','LageCap','MidCap','SmallCap','TinyCap','Industry',
            '2015_01_CLOSE','2015_02_CLOSE','2015_03_CLOSE','2015_04_CLOSE','2015_05_CLOSE','2015_06_CLOSE','2015_07_CLOSE','2015_08_CLOSE','2015_09_CLOSE','2015_10_CLOSE','2015_11_CLOSE','2015_12_CLOSE',
            '2016']

list_2017 = ['Tikr',
            'Sales_2012','Sales_2013','Sales_2014','Sales_2015','Sales_2016',
            'Sales_2013_YoY','Sales_2014_YoY','Sales_2015_YoY','Sales_2016_YoY',
            'Revenue_2012','Revenue_2013','Revenue_2014','Revenue_2015','Revenue_2016',
            'Revenue_2013_YoY','Revenue_2014_YoY','Revenue_2015_YoY','Revenue_2016_YoY',
            'Expenses_2012','Expenses_2013','Expenses_2014','Expenses_2015','Expenses_2016',
            'Expenses_2013_YoY','Expenses_2014_YoY','Expenses_2015_YoY','Expenses_2016_YoY',
            'Net_Profit_2012','Net_Profit_2013','Net_Profit_2014','Net_Profit_2015','Net_Profit_2016',
            'Net_Profit_2013_YoY','Net_Profit_2014_YoY','Net_Profit_2015_YoY','Net_Profit_2016_YoY',
            'Operating_Profit_2012','Operating_Profit_2013','Operating_Profit_2014','Operating_Profit_2015','Operating_Profit_2016',
            'Operating_Profit_2013_YoY','Operating_Profit_2014_YoY','Operating_Profit_2015_YoY','Operating_Profit_2016_YoY',
            'EPS_in_Rs_2012','EPS_in_Rs_2013','EPS_in_Rs_2014','EPS_in_Rs_2015','EPS_in_Rs_2016',
            'EPS_in_Rs_2013_YoY','EPS_in_Rs_2014_YoY','EPS_in_Rs_2015_YoY','EPS_in_Rs_2016_YoY',
            'Tax__2012','Tax__2013','Tax__2014','Tax__2015','Tax__2016',
            'Tax__2013_YoY','Tax__2014_YoY','Tax__2015_YoY','Tax__2016_YoY',
            'OPM__2012','OPM__2013','OPM__2014','OPM__2015','OPM__2016',
            'OPM__2013_YoY','OPM__2014_YoY','OPM__2015_YoY','OPM__2016_YoY',
            'Other_Assets_2012','Other_Assets_2013','Other_Assets_2014','Other_Assets_2015','Other_Assets_2016',
            'Other_Assets_2013_YoY','Other_Assets_2014_YoY','Other_Assets_2015_YoY','Other_Assets_2016_YoY',
            'Total_Liabilities_2012','Total_Liabilities_2013','Total_Liabilities_2014','Total_Liabilities_2015','Total_Liabilities_2016',
            'Total_Liabilities_2013_YoY','Total_Liabilities_2014_YoY','Total_Liabilities_2015_YoY','Total_Liabilities_2016_YoY',
            'Share_Capital_2012','Share_Capital_2013','Share_Capital_2014','Share_Capital_2015','Share_Capital_2016',
            'Share_Capital_2013_YoY','Share_Capital_2014_YoY','Share_Capital_2015_YoY','Share_Capital_2016_YoY',
            'Investments_2012','Investments_2013','Investments_2014','Investments_2015','Investments_2016',
            'Investments_2013_YoY','Investments_2014_YoY','Investments_2015_YoY','Investments_2016_YoY',
            'Borrowings_2012','Borrowings_2013','Borrowings_2014','Borrowings_2015','Borrowings_2016',
            'Borrowings_2013_YoY','Borrowings_2014_YoY','Borrowings_2015_YoY','Borrowings_2016_YoY',
            'Reserves_2012','Reserves_2013','Reserves_2014','Reserves_2015','Reserves_2016',
            'Reserves_2013_YoY','Reserves_2014_YoY','Reserves_2015_YoY','Reserves_2016_YoY',
            'Total_Assets_2012','Total_Assets_2013','Total_Assets_2014','Total_Assets_2015','Total_Assets_2016',
            'Total_Assets_2013_YoY','Total_Assets_2014_YoY','Total_Assets_2015_YoY','Total_Assets_2016_YoY',
            'Fixed_Assets_2012','Fixed_Assets_2013','Fixed_Assets_2014','Fixed_Assets_2015','Fixed_Assets_2016',
            'Fixed_Assets_2013_YoY','Fixed_Assets_2014_YoY','Fixed_Assets_2015_YoY','Fixed_Assets_2016_YoY',
            'Other_Liabilities_2012','Other_Liabilities_2013','Other_Liabilities_2014','Other_Liabilities_2015','Other_Liabilities_2016',
            'Other_Liabilities_2013_YoY','Other_Liabilities_2014_YoY','Other_Liabilities_2015_YoY','Other_Liabilities_2016_YoY',
            'CWIP_2012','CWIP_2013','CWIP_2014','CWIP_2015','CWIP_2016',
            'CWIP_2013_YoY','CWIP_2014_YoY','CWIP_2015_YoY','CWIP_2016_YoY',
            'Cash_from_Investing_Activity_2012','Cash_from_Investing_Activity_2013','Cash_from_Investing_Activity_2014','Cash_from_Investing_Activity_2015','Cash_from_Investing_Activity_2016',
            'Cash_from_Investing_Activity_2013_YoY','Cash_from_Investing_Activity_2014_YoY','Cash_from_Investing_Activity_2015_YoY','Cash_from_Investing_Activity_2016_YoY',
            'Cash_from_Operating_Activity_2012','Cash_from_Operating_Activity_2013','Cash_from_Operating_Activity_2014','Cash_from_Operating_Activity_2015','Cash_from_Operating_Activity_2016',
            'Cash_from_Operating_Activity_2013_YoY','Cash_from_Operating_Activity_2014_YoY','Cash_from_Operating_Activity_2015_YoY','Cash_from_Operating_Activity_2016_YoY',
            'Cash_from_Financing_Activity_2012','Cash_from_Financing_Activity_2013','Cash_from_Financing_Activity_2014','Cash_from_Financing_Activity_2015','Cash_from_Financing_Activity_2016',
            'Cash_from_Financing_Activity_2013_YoY','Cash_from_Financing_Activity_2014_YoY','Cash_from_Financing_Activity_2015_YoY','Cash_from_Financing_Activity_2016_YoY',
            'Net_Cash_Flow_2012','Net_Cash_Flow_2013','Net_Cash_Flow_2014','Net_Cash_Flow_2015','Net_Cash_Flow_2016',
            'Net_Cash_Flow_2013_YoY','Net_Cash_Flow_2014_YoY','Net_Cash_Flow_2015_YoY','Net_Cash_Flow_2016_YoY',
            'Share_Capital_%_2012','Share_Capital_%_2013','Share_Capital_%_2014','Share_Capital_%_2015','Share_Capital_%_2016',
            'Reserves_%_2012','Reserves_%_2013','Reserves_%_2014','Reserves_%_2015','Reserves_%_2016',
            'Borrowings_%_2012','Borrowings_%_2013','Borrowings_%_2014','Borrowings_%_2015','Borrowings_%_2016',
            'Other_Liabilities_%_2012','Other_Liabilities_%_2013','Other_Liabilities_%_2014','Other_Liabilities_%_2015','Other_Liabilities_%_2016',
            'Fixed_Assets_%_2012','Fixed_Assets_%_2013','Fixed_Assets_%_2014','Fixed_Assets_%_2015','Fixed_Assets_%_2016',
            'CWIP_%_2012','CWIP_%_2013','CWIP_%_2014','CWIP_%_2015','CWIP_%_2016',
            'Investments_%_2012','Investments_%_2013','Investments_%_2014','Investments_%_2015','Investments_%_2016',
            'Other_Assets_%_2012','Other_Assets_%_2013','Other_Assets_%_2014','Other_Assets_%_2015','Other_Assets_%_2016',
            'MarketCap','CapFlag','LageCap','MidCap','SmallCap','TinyCap','Industry',
            '2016_01_CLOSE','2016_02_CLOSE','2016_03_CLOSE','2016_04_CLOSE','2016_05_CLOSE','2016_06_CLOSE','2016_07_CLOSE','2016_08_CLOSE','2016_09_CLOSE','2016_10_CLOSE','2016_11_CLOSE','2016_12_CLOSE',
            '2017']
    
list_2018 = ['Tikr',
            'Sales_2013','Sales_2014','Sales_2015','Sales_2016','Sales_2017',
            'Sales_2014_YoY','Sales_2015_YoY','Sales_2016_YoY','Sales_2017_YoY',
            'Revenue_2013','Revenue_2014','Revenue_2015','Revenue_2016','Revenue_2017',
            'Revenue_2014_YoY','Revenue_2015_YoY','Revenue_2016_YoY','Revenue_2017_YoY',
            'Expenses_2013','Expenses_2014','Expenses_2015','Expenses_2016','Expenses_2017',
            'Expenses_2014_YoY','Expenses_2015_YoY','Expenses_2016_YoY','Expenses_2017_YoY',
            'Net_Profit_2013','Net_Profit_2014','Net_Profit_2015','Net_Profit_2016','Net_Profit_2017',
            'Net_Profit_2014_YoY','Net_Profit_2015_YoY','Net_Profit_2016_YoY','Net_Profit_2017_YoY',
            'Operating_Profit_2013','Operating_Profit_2014','Operating_Profit_2015','Operating_Profit_2016','Operating_Profit_2017',
            'Operating_Profit_2014_YoY','Operating_Profit_2015_YoY','Operating_Profit_2016_YoY','Operating_Profit_2017_YoY',
            'EPS_in_Rs_2013','EPS_in_Rs_2014','EPS_in_Rs_2015','EPS_in_Rs_2016','EPS_in_Rs_2017',
            'EPS_in_Rs_2014_YoY','EPS_in_Rs_2015_YoY','EPS_in_Rs_2016_YoY','EPS_in_Rs_2017_YoY',
            'Tax__2013','Tax__2014','Tax__2015','Tax__2016','Tax__2017',
            'Tax__2014_YoY','Tax__2015_YoY','Tax__2016_YoY','Tax__2017_YoY',
            'OPM__2013','OPM__2014','OPM__2015','OPM__2016','OPM__2017',
            'OPM__2014_YoY','OPM__2015_YoY','OPM__2016_YoY','OPM__2017_YoY',
            'Other_Assets_2013','Other_Assets_2014','Other_Assets_2015','Other_Assets_2016','Other_Assets_2017',
            'Other_Assets_2014_YoY','Other_Assets_2015_YoY','Other_Assets_2016_YoY','Other_Assets_2017_YoY',
            'Total_Liabilities_2013','Total_Liabilities_2014','Total_Liabilities_2015','Total_Liabilities_2016','Total_Liabilities_2017',
            'Total_Liabilities_2014_YoY','Total_Liabilities_2015_YoY','Total_Liabilities_2016_YoY','Total_Liabilities_2017_YoY',
            'Share_Capital_2013','Share_Capital_2014','Share_Capital_2015','Share_Capital_2016','Share_Capital_2017',
            'Share_Capital_2014_YoY','Share_Capital_2015_YoY','Share_Capital_2016_YoY','Share_Capital_2017_YoY',
            'Investments_2013','Investments_2014','Investments_2015','Investments_2016','Investments_2017',
            'Investments_2014_YoY','Investments_2015_YoY','Investments_2016_YoY','Investments_2017_YoY',
            'Borrowings_2013','Borrowings_2014','Borrowings_2015','Borrowings_2016','Borrowings_2017',
            'Borrowings_2014_YoY','Borrowings_2015_YoY','Borrowings_2016_YoY','Borrowings_2017_YoY',
            'Reserves_2013','Reserves_2014','Reserves_2015','Reserves_2016','Reserves_2017',
            'Reserves_2014_YoY','Reserves_2015_YoY','Reserves_2016_YoY','Reserves_2017_YoY',
            'Total_Assets_2013','Total_Assets_2014','Total_Assets_2015','Total_Assets_2016','Total_Assets_2017',
            'Total_Assets_2014_YoY','Total_Assets_2015_YoY','Total_Assets_2016_YoY','Total_Assets_2017_YoY',
            'Fixed_Assets_2013','Fixed_Assets_2014','Fixed_Assets_2015','Fixed_Assets_2016','Fixed_Assets_2017',
            'Fixed_Assets_2014_YoY','Fixed_Assets_2015_YoY','Fixed_Assets_2016_YoY','Fixed_Assets_2017_YoY',
            'Other_Liabilities_2013','Other_Liabilities_2014','Other_Liabilities_2015','Other_Liabilities_2016','Other_Liabilities_2017',
            'Other_Liabilities_2014_YoY','Other_Liabilities_2015_YoY','Other_Liabilities_2016_YoY','Other_Liabilities_2017_YoY',
            'CWIP_2013','CWIP_2014','CWIP_2015','CWIP_2016','CWIP_2017',
            'CWIP_2014_YoY','CWIP_2015_YoY','CWIP_2016_YoY','CWIP_2017_YoY',
            'Cash_from_Investing_Activity_2013','Cash_from_Investing_Activity_2014','Cash_from_Investing_Activity_2015','Cash_from_Investing_Activity_2016','Cash_from_Investing_Activity_2017',
            'Cash_from_Investing_Activity_2014_YoY','Cash_from_Investing_Activity_2015_YoY','Cash_from_Investing_Activity_2016_YoY','Cash_from_Investing_Activity_2017_YoY',
            'Cash_from_Operating_Activity_2013','Cash_from_Operating_Activity_2014','Cash_from_Operating_Activity_2015','Cash_from_Operating_Activity_2016','Cash_from_Operating_Activity_2017',
            'Cash_from_Operating_Activity_2014_YoY','Cash_from_Operating_Activity_2015_YoY','Cash_from_Operating_Activity_2016_YoY','Cash_from_Operating_Activity_2017_YoY',
            'Cash_from_Financing_Activity_2013','Cash_from_Financing_Activity_2014','Cash_from_Financing_Activity_2015','Cash_from_Financing_Activity_2016','Cash_from_Financing_Activity_2017',
            'Cash_from_Financing_Activity_2014_YoY','Cash_from_Financing_Activity_2015_YoY','Cash_from_Financing_Activity_2016_YoY','Cash_from_Financing_Activity_2017_YoY',
            'Net_Cash_Flow_2013','Net_Cash_Flow_2014','Net_Cash_Flow_2015','Net_Cash_Flow_2016','Net_Cash_Flow_2017',
            'Net_Cash_Flow_2014_YoY','Net_Cash_Flow_2015_YoY','Net_Cash_Flow_2016_YoY','Net_Cash_Flow_2017_YoY',
            'Share_Capital_%_2013','Share_Capital_%_2014','Share_Capital_%_2015','Share_Capital_%_2016','Share_Capital_%_2017',
            'Reserves_%_2013','Reserves_%_2014','Reserves_%_2015','Reserves_%_2016','Reserves_%_2017',
            'Borrowings_%_2013','Borrowings_%_2014','Borrowings_%_2015','Borrowings_%_2016','Borrowings_%_2017',
            'Other_Liabilities_%_2013','Other_Liabilities_%_2014','Other_Liabilities_%_2015','Other_Liabilities_%_2016','Other_Liabilities_%_2017',
            'Fixed_Assets_%_2013','Fixed_Assets_%_2014','Fixed_Assets_%_2015','Fixed_Assets_%_2016','Fixed_Assets_%_2017',
            'CWIP_%_2013','CWIP_%_2014','CWIP_%_2015','CWIP_%_2016','CWIP_%_2017',
            'Investments_%_2013','Investments_%_2014','Investments_%_2015','Investments_%_2016','Investments_%_2017',
            'Other_Assets_%_2013','Other_Assets_%_2014','Other_Assets_%_2015','Other_Assets_%_2016','Other_Assets_%_2017',
            'MarketCap','CapFlag','LageCap','MidCap','SmallCap','TinyCap','Industry',
            '2017_01_CLOSE','2017_02_CLOSE','2017_03_CLOSE','2017_04_CLOSE','2017_05_CLOSE','2017_06_CLOSE','2017_07_CLOSE','2017_08_CLOSE','2017_09_CLOSE','2017_10_CLOSE','2017_11_CLOSE','2017_12_CLOSE',
            '2018']


list_2019 = ['Tikr',
            'Sales_2014','Sales_2015','Sales_2016','Sales_2017','Sales_2018',
            'Sales_2015_YoY','Sales_2016_YoY','Sales_2017_YoY','Sales_2018_YoY',
            'Revenue_2014','Revenue_2015','Revenue_2016','Revenue_2017','Revenue_2018',
            'Revenue_2015_YoY','Revenue_2016_YoY','Revenue_2017_YoY','Revenue_2018_YoY',
            'Expenses_2014','Expenses_2015','Expenses_2016','Expenses_2017','Expenses_2018',
            'Expenses_2015_YoY','Expenses_2016_YoY','Expenses_2017_YoY','Expenses_2018_YoY',
            'Net_Profit_2014','Net_Profit_2015','Net_Profit_2016','Net_Profit_2017','Net_Profit_2018',
            'Net_Profit_2015_YoY','Net_Profit_2016_YoY','Net_Profit_2017_YoY','Net_Profit_2018_YoY',
            'Operating_Profit_2014','Operating_Profit_2015','Operating_Profit_2016','Operating_Profit_2017','Operating_Profit_2018',
            'Operating_Profit_2015_YoY','Operating_Profit_2016_YoY','Operating_Profit_2017_YoY','Operating_Profit_2018_YoY',
            'EPS_in_Rs_2014','EPS_in_Rs_2015','EPS_in_Rs_2016','EPS_in_Rs_2017','EPS_in_Rs_2018',
            'EPS_in_Rs_2015_YoY','EPS_in_Rs_2016_YoY','EPS_in_Rs_2017_YoY','EPS_in_Rs_2018_YoY',
            'Tax__2014','Tax__2015','Tax__2016','Tax__2017','Tax__2018',
            'Tax__2015_YoY','Tax__2016_YoY','Tax__2017_YoY','Tax__2018_YoY',
            'OPM__2014','OPM__2015','OPM__2016','OPM__2017','OPM__2018',
            'OPM__2015_YoY','OPM__2016_YoY','OPM__2017_YoY','OPM__2018_YoY',
            'Other_Assets_2014','Other_Assets_2015','Other_Assets_2016','Other_Assets_2017','Other_Assets_2018',
            'Other_Assets_2015_YoY','Other_Assets_2016_YoY','Other_Assets_2017_YoY','Other_Assets_2018_YoY',
            'Total_Liabilities_2014','Total_Liabilities_2015','Total_Liabilities_2016','Total_Liabilities_2017','Total_Liabilities_2018',
            'Total_Liabilities_2015_YoY','Total_Liabilities_2016_YoY','Total_Liabilities_2017_YoY','Total_Liabilities_2018_YoY',
            'Share_Capital_2014','Share_Capital_2015','Share_Capital_2016','Share_Capital_2017','Share_Capital_2018',
            'Share_Capital_2015_YoY','Share_Capital_2016_YoY','Share_Capital_2017_YoY','Share_Capital_2018_YoY',
            'Investments_2014','Investments_2015','Investments_2016','Investments_2017','Investments_2018',
            'Investments_2015_YoY','Investments_2016_YoY','Investments_2017_YoY','Investments_2018_YoY',
            'Borrowings_2014','Borrowings_2015','Borrowings_2016','Borrowings_2017','Borrowings_2018',
            'Borrowings_2015_YoY','Borrowings_2016_YoY','Borrowings_2017_YoY','Borrowings_2018_YoY',
            'Reserves_2014','Reserves_2015','Reserves_2016','Reserves_2017','Reserves_2018',
            'Reserves_2015_YoY','Reserves_2016_YoY','Reserves_2017_YoY','Reserves_2018_YoY',
            'Total_Assets_2014','Total_Assets_2015','Total_Assets_2016','Total_Assets_2017','Total_Assets_2018',
            'Total_Assets_2015_YoY','Total_Assets_2016_YoY','Total_Assets_2017_YoY','Total_Assets_2018_YoY',
            'Fixed_Assets_2014','Fixed_Assets_2015','Fixed_Assets_2016','Fixed_Assets_2017','Fixed_Assets_2018',
            'Fixed_Assets_2015_YoY','Fixed_Assets_2016_YoY','Fixed_Assets_2017_YoY','Fixed_Assets_2018_YoY',
            'Other_Liabilities_2014','Other_Liabilities_2015','Other_Liabilities_2016','Other_Liabilities_2017','Other_Liabilities_2018',
            'Other_Liabilities_2015_YoY','Other_Liabilities_2016_YoY','Other_Liabilities_2017_YoY','Other_Liabilities_2018_YoY',
            'CWIP_2014','CWIP_2015','CWIP_2016','CWIP_2017','CWIP_2018',
            'CWIP_2015_YoY','CWIP_2016_YoY','CWIP_2017_YoY','CWIP_2018_YoY',
            'Cash_from_Investing_Activity_2014','Cash_from_Investing_Activity_2015','Cash_from_Investing_Activity_2016','Cash_from_Investing_Activity_2017','Cash_from_Investing_Activity_2018',
            'Cash_from_Investing_Activity_2015_YoY','Cash_from_Investing_Activity_2016_YoY','Cash_from_Investing_Activity_2017_YoY','Cash_from_Investing_Activity_2018_YoY',
            'Cash_from_Operating_Activity_2014','Cash_from_Operating_Activity_2015','Cash_from_Operating_Activity_2016','Cash_from_Operating_Activity_2017','Cash_from_Operating_Activity_2018',
            'Cash_from_Operating_Activity_2015_YoY','Cash_from_Operating_Activity_2016_YoY','Cash_from_Operating_Activity_2017_YoY','Cash_from_Operating_Activity_2018_YoY',
            'Cash_from_Financing_Activity_2014','Cash_from_Financing_Activity_2015','Cash_from_Financing_Activity_2016','Cash_from_Financing_Activity_2017','Cash_from_Financing_Activity_2018',
            'Cash_from_Financing_Activity_2015_YoY','Cash_from_Financing_Activity_2016_YoY','Cash_from_Financing_Activity_2017_YoY','Cash_from_Financing_Activity_2018_YoY',
            'Net_Cash_Flow_2014','Net_Cash_Flow_2015','Net_Cash_Flow_2016','Net_Cash_Flow_2017','Net_Cash_Flow_2018',
            'Net_Cash_Flow_2015_YoY','Net_Cash_Flow_2016_YoY','Net_Cash_Flow_2017_YoY','Net_Cash_Flow_2018_YoY',
            'Share_Capital_%_2014','Share_Capital_%_2015','Share_Capital_%_2016','Share_Capital_%_2017','Share_Capital_%_2018',
            'Reserves_%_2014','Reserves_%_2015','Reserves_%_2016','Reserves_%_2017','Reserves_%_2018',
            'Borrowings_%_2014','Borrowings_%_2015','Borrowings_%_2016','Borrowings_%_2017','Borrowings_%_2018',
            'Other_Liabilities_%_2014','Other_Liabilities_%_2015','Other_Liabilities_%_2016','Other_Liabilities_%_2017','Other_Liabilities_%_2018',
            'Fixed_Assets_%_2014','Fixed_Assets_%_2015','Fixed_Assets_%_2016','Fixed_Assets_%_2017','Fixed_Assets_%_2018',
            'CWIP_%_2014','CWIP_%_2015','CWIP_%_2016','CWIP_%_2017','CWIP_%_2018',
            'Investments_%_2014','Investments_%_2015','Investments_%_2016','Investments_%_2017','Investments_%_2018',
            'Other_Assets_%_2014','Other_Assets_%_2015','Other_Assets_%_2016','Other_Assets_%_2017','Other_Assets_%_2018',
            'MarketCap','CapFlag','LageCap','MidCap','SmallCap','TinyCap','Industry',
            '2018_01_CLOSE','2018_02_CLOSE','2018_03_CLOSE','2018_04_CLOSE','2018_05_CLOSE','2018_06_CLOSE','2018_07_CLOSE','2018_08_CLOSE','2018_09_CLOSE','2018_10_CLOSE','2018_11_CLOSE','2018_12_CLOSE',
            '2019']

# list_2020 = ['Tikr',        
#             'Sales_2015','Sales_2016','Sales_2017','Sales_2018','Sales_2019',
#             'Sales_2016_YoY','Sales_2017_YoY','Sales_2018_YoY','Sales_2019_YoY',
#             'Revenue_2015','Revenue_2016','Revenue_2017','Revenue_2018','Revenue_2019',
#             'Revenue_2016_YoY','Revenue_2017_YoY','Revenue_2018_YoY','Revenue_2019_YoY',
#             'Expenses_2015','Expenses_2016','Expenses_2017','Expenses_2018','Expenses_2019',
#             'Expenses_2016_YoY','Expenses_2017_YoY','Expenses_2018_YoY','Expenses_2019_YoY',
#             'Net_Profit_2015','Net_Profit_2016','Net_Profit_2017','Net_Profit_2018','Net_Profit_2019',
#             'Net_Profit_2016_YoY','Net_Profit_2017_YoY','Net_Profit_2018_YoY','Net_Profit_2019_YoY',
#             'Operating_Profit_2015','Operating_Profit_2016','Operating_Profit_2017','Operating_Profit_2018','Operating_Profit_2019',
#             'Operating_Profit_2016_YoY','Operating_Profit_2017_YoY','Operating_Profit_2018_YoY','Operating_Profit_2019_YoY',
#             'EPS_in_Rs_2015','EPS_in_Rs_2016','EPS_in_Rs_2017','EPS_in_Rs_2018','EPS_in_Rs_2019',
#             'EPS_in_Rs_2016_YoY','EPS_in_Rs_2017_YoY','EPS_in_Rs_2018_YoY','EPS_in_Rs_2019_YoY',
#             'Tax__2015','Tax__2016','Tax__2017','Tax__2018','Tax__2019',
#             'Tax__2016_YoY','Tax__2017_YoY','Tax__2018_YoY','Tax__2019_YoY',
#             'OPM__2015','OPM__2016','OPM__2017','OPM__2018','OPM__2019',
#             'OPM__2016_YoY','OPM__2017_YoY','OPM__2018_YoY','OPM__2019_YoY',
#             'Other_Assets_2015','Other_Assets_2016','Other_Assets_2017','Other_Assets_2018','Other_Assets_2019',
#             'Other_Assets_2016_YoY','Other_Assets_2017_YoY','Other_Assets_2018_YoY','Other_Assets_2019_YoY',
#             'Total_Liabilities_2015','Total_Liabilities_2016','Total_Liabilities_2017','Total_Liabilities_2018','Total_Liabilities_2019',
#             'Total_Liabilities_2016_YoY','Total_Liabilities_2017_YoY','Total_Liabilities_2018_YoY','Total_Liabilities_2019_YoY',
#             'Share_Capital_2015','Share_Capital_2016','Share_Capital_2017','Share_Capital_2018','Share_Capital_2019',
#             'Share_Capital_2016_YoY','Share_Capital_2017_YoY','Share_Capital_2018_YoY','Share_Capital_2019_YoY',
#             'Investments_2015','Investments_2016','Investments_2017','Investments_2018','Investments_2019',
#             'Investments_2016_YoY','Investments_2017_YoY','Investments_2018_YoY','Investments_2019_YoY',
#             'Borrowings_2015','Borrowings_2016','Borrowings_2017','Borrowings_2018','Borrowings_2019',
#             'Borrowings_2016_YoY','Borrowings_2017_YoY','Borrowings_2018_YoY','Borrowings_2019_YoY',
#             'Reserves_2015','Reserves_2016','Reserves_2017','Reserves_2018','Reserves_2019',
#             'Reserves_2016_YoY','Reserves_2017_YoY','Reserves_2018_YoY','Reserves_2019_YoY',
#             'Total_Assets_2015','Total_Assets_2016','Total_Assets_2017','Total_Assets_2018','Total_Assets_2019',
#             'Total_Assets_2016_YoY','Total_Assets_2017_YoY','Total_Assets_2018_YoY','Total_Assets_2019_YoY',
#             'Fixed_Assets_2015','Fixed_Assets_2016','Fixed_Assets_2017','Fixed_Assets_2018','Fixed_Assets_2019',
#             'Fixed_Assets_2016_YoY','Fixed_Assets_2017_YoY','Fixed_Assets_2018_YoY','Fixed_Assets_2019_YoY',
#             'Other_Liabilities_2015','Other_Liabilities_2016','Other_Liabilities_2017','Other_Liabilities_2018','Other_Liabilities_2019',
#             'Other_Liabilities_2016_YoY','Other_Liabilities_2017_YoY','Other_Liabilities_2018_YoY','Other_Liabilities_2019_YoY',
#             'CWIP_2015','CWIP_2016','CWIP_2017','CWIP_2018','CWIP_2019',
#             'CWIP_2016_YoY','CWIP_2017_YoY','CWIP_2018_YoY','CWIP_2019_YoY',
#             'Cash_from_Investing_Activity_2015','Cash_from_Investing_Activity_2016','Cash_from_Investing_Activity_2017','Cash_from_Investing_Activity_2018','Cash_from_Investing_Activity_2019',
#             'Cash_from_Investing_Activity_2016_YoY','Cash_from_Investing_Activity_2017_YoY','Cash_from_Investing_Activity_2018_YoY','Cash_from_Investing_Activity_2019_YoY',
#             'Cash_from_Operating_Activity_2015','Cash_from_Operating_Activity_2016','Cash_from_Operating_Activity_2017','Cash_from_Operating_Activity_2018','Cash_from_Operating_Activity_2019',
#             'Cash_from_Operating_Activity_2016_YoY','Cash_from_Operating_Activity_2017_YoY','Cash_from_Operating_Activity_2018_YoY','Cash_from_Operating_Activity_2019_YoY',
#             'Cash_from_Financing_Activity_2015','Cash_from_Financing_Activity_2016','Cash_from_Financing_Activity_2017','Cash_from_Financing_Activity_2018','Cash_from_Financing_Activity_2019',
#             'Cash_from_Financing_Activity_2016_YoY','Cash_from_Financing_Activity_2017_YoY','Cash_from_Financing_Activity_2018_YoY','Cash_from_Financing_Activity_2019_YoY',
#             'Net_Cash_Flow_2015','Net_Cash_Flow_2016','Net_Cash_Flow_2017','Net_Cash_Flow_2018','Net_Cash_Flow_2019',
#             'Net_Cash_Flow_2016_YoY','Net_Cash_Flow_2017_YoY','Net_Cash_Flow_2018_YoY','Net_Cash_Flow_2019_YoY',
#             'Share_Capital_%_2015','Share_Capital_%_2016','Share_Capital_%_2017','Share_Capital_%_2018','Share_Capital_%_2019',
#             'Reserves_%_2015','Reserves_%_2016','Reserves_%_2017','Reserves_%_2018','Reserves_%_2019',
#             'Borrowings_%_2015','Borrowings_%_2016','Borrowings_%_2017','Borrowings_%_2018','Borrowings_%_2019',
#             'Other_Liabilities_%_2015','Other_Liabilities_%_2016','Other_Liabilities_%_2017','Other_Liabilities_%_2018','Other_Liabilities_%_2019',
#             'Fixed_Assets_%_2015','Fixed_Assets_%_2016','Fixed_Assets_%_2017','Fixed_Assets_%_2018','Fixed_Assets_%_2019',
#             'CWIP_%_2015','CWIP_%_2016','CWIP_%_2017','CWIP_%_2018','CWIP_%_2019',
#             'Investments_%_2015','Investments_%_2016','Investments_%_2017','Investments_%_2018','Investments_%_2019',
#             'Other_Assets_%_2015','Other_Assets_%_2016','Other_Assets_%_2017','Other_Assets_%_2018','Other_Assets_%_2019',
#             'MarketCap','CapFlag','LageCap','MidCap','SmallCap','TinyCap','Industry',
#             '2019_01_CLOSE','2019_02_CLOSE','2019_03_CLOSE','2019_04_CLOSE','2019_05_CLOSE','2019_06_CLOSE','2019_07_CLOSE','2019_08_CLOSE','2019_09_CLOSE','2019_10_CLOSE','2019_11_CLOSE','2019_12_CLOSE',
#             '2020']

# list_2021 = ['Tikr',
#             'Sales_2016','Sales_2017','Sales_2018','Sales_2019','Sales_2020',
#             'Sales_2017_YoY','Sales_2018_YoY','Sales_2019_YoY','Sales_2020_YoY',
#             'Revenue_2016','Revenue_2017','Revenue_2018','Revenue_2019','Revenue_2020',
#             'Revenue_2017_YoY','Revenue_2018_YoY','Revenue_2019_YoY','Revenue_2020_YoY',
#             'Expenses_2016','Expenses_2017','Expenses_2018','Expenses_2019','Expenses_2020',
#             'Expenses_2017_YoY','Expenses_2018_YoY','Expenses_2019_YoY','Expenses_2020_YoY',
#             'Net_Profit_2016','Net_Profit_2017','Net_Profit_2018','Net_Profit_2019','Net_Profit_2020',
#             'Net_Profit_2017_YoY','Net_Profit_2018_YoY','Net_Profit_2019_YoY','Net_Profit_2020_YoY',
#             'Operating_Profit_2016','Operating_Profit_2017','Operating_Profit_2018','Operating_Profit_2019','Operating_Profit_2020',
#             'Operating_Profit_2017_YoY','Operating_Profit_2018_YoY','Operating_Profit_2019_YoY','Operating_Profit_2020_YoY',
#             'EPS_in_Rs_2016','EPS_in_Rs_2017','EPS_in_Rs_2018','EPS_in_Rs_2019','EPS_in_Rs_2020',
#             'EPS_in_Rs_2017_YoY','EPS_in_Rs_2018_YoY','EPS_in_Rs_2019_YoY','EPS_in_Rs_2020_YoY',
#             'Tax__2016','Tax__2017','Tax__2018','Tax__2019','Tax__2020',
#             'Tax__2017_YoY','Tax__2018_YoY','Tax__2019_YoY','Tax__2020_YoY',
#             'OPM__2016','OPM__2017','OPM__2018','OPM__2019','OPM__2020',
#             'OPM__2017_YoY','OPM__2018_YoY','OPM__2019_YoY','OPM__2020_YoY',
#             'Other_Assets_2016','Other_Assets_2017','Other_Assets_2018','Other_Assets_2019','Other_Assets_2020',
#             'Other_Assets_2017_YoY','Other_Assets_2018_YoY','Other_Assets_2019_YoY','Other_Assets_2020_YoY',
#             'Total_Liabilities_2016','Total_Liabilities_2017','Total_Liabilities_2018','Total_Liabilities_2019','Total_Liabilities_2020',
#             'Total_Liabilities_2017_YoY','Total_Liabilities_2018_YoY','Total_Liabilities_2019_YoY','Total_Liabilities_2020_YoY',
#             'Share_Capital_2016','Share_Capital_2017','Share_Capital_2018','Share_Capital_2019','Share_Capital_2020',
#             'Share_Capital_2017_YoY','Share_Capital_2018_YoY','Share_Capital_2019_YoY','Share_Capital_2020_YoY',
#             'Investments_2016','Investments_2017','Investments_2018','Investments_2019','Investments_2020',
#             'Investments_2017_YoY','Investments_2018_YoY','Investments_2019_YoY','Investments_2020_YoY',
#             'Borrowings_2016','Borrowings_2017','Borrowings_2018','Borrowings_2019','Borrowings_2020',
#             'Borrowings_2017_YoY','Borrowings_2018_YoY','Borrowings_2019_YoY','Borrowings_2020_YoY',
#             'Reserves_2016','Reserves_2017','Reserves_2018','Reserves_2019','Reserves_2020',
#             'Reserves_2017_YoY','Reserves_2018_YoY','Reserves_2019_YoY','Reserves_2020_YoY',
#             'Total_Assets_2016','Total_Assets_2017','Total_Assets_2018','Total_Assets_2019','Total_Assets_2020',
#             'Total_Assets_2017_YoY','Total_Assets_2018_YoY','Total_Assets_2019_YoY','Total_Assets_2020_YoY',
#             'Fixed_Assets_2016','Fixed_Assets_2017','Fixed_Assets_2018','Fixed_Assets_2019','Fixed_Assets_2020',
#             'Fixed_Assets_2017_YoY','Fixed_Assets_2018_YoY','Fixed_Assets_2019_YoY','Fixed_Assets_2020_YoY',
#             'Other_Liabilities_2016','Other_Liabilities_2017','Other_Liabilities_2018','Other_Liabilities_2019','Other_Liabilities_2020',
#             'Other_Liabilities_2017_YoY','Other_Liabilities_2018_YoY','Other_Liabilities_2019_YoY','Other_Liabilities_2020_YoY',
#             'CWIP_2016','CWIP_2017','CWIP_2018','CWIP_2019','CWIP_2020',
#             'CWIP_2017_YoY','CWIP_2018_YoY','CWIP_2019_YoY','CWIP_2020_YoY',
#             'Cash_from_Investing_Activity_2016','Cash_from_Investing_Activity_2017','Cash_from_Investing_Activity_2018','Cash_from_Investing_Activity_2019','Cash_from_Investing_Activity_2020',
#             'Cash_from_Investing_Activity_2017_YoY','Cash_from_Investing_Activity_2018_YoY','Cash_from_Investing_Activity_2019_YoY','Cash_from_Investing_Activity_2020_YoY',
#             'Cash_from_Operating_Activity_2016','Cash_from_Operating_Activity_2017','Cash_from_Operating_Activity_2018','Cash_from_Operating_Activity_2019','Cash_from_Operating_Activity_2020',
#             'Cash_from_Operating_Activity_2017_YoY','Cash_from_Operating_Activity_2018_YoY','Cash_from_Operating_Activity_2019_YoY','Cash_from_Operating_Activity_2020_YoY',
#             'Cash_from_Financing_Activity_2016','Cash_from_Financing_Activity_2017','Cash_from_Financing_Activity_2018','Cash_from_Financing_Activity_2019','Cash_from_Financing_Activity_2020',
#             'Cash_from_Financing_Activity_2017_YoY','Cash_from_Financing_Activity_2018_YoY','Cash_from_Financing_Activity_2019_YoY','Cash_from_Financing_Activity_2020_YoY',
#             'Net_Cash_Flow_2016','Net_Cash_Flow_2017','Net_Cash_Flow_2018','Net_Cash_Flow_2019','Net_Cash_Flow_2020',
#             'Net_Cash_Flow_2017_YoY','Net_Cash_Flow_2018_YoY','Net_Cash_Flow_2019_YoY','Net_Cash_Flow_2020_YoY',
#             'Share_Capital_%_2016','Share_Capital_%_2017','Share_Capital_%_2018','Share_Capital_%_2019','Share_Capital_%_2020',
#             'Reserves_%_2016','Reserves_%_2017','Reserves_%_2018','Reserves_%_2019','Reserves_%_2020',
#             'Borrowings_%_2016','Borrowings_%_2017','Borrowings_%_2018','Borrowings_%_2019','Borrowings_%_2020',
#             'Other_Liabilities_%_2016','Other_Liabilities_%_2017','Other_Liabilities_%_2018','Other_Liabilities_%_2019','Other_Liabilities_%_2020',
#             'Fixed_Assets_%_2016','Fixed_Assets_%_2017','Fixed_Assets_%_2018','Fixed_Assets_%_2019','Fixed_Assets_%_2020',
#             'CWIP_%_2016','CWIP_%_2017','CWIP_%_2018','CWIP_%_2019','CWIP_%_2020',
#             'Investments_%_2016','Investments_%_2017','Investments_%_2018','Investments_%_2019','Investments_%_2020',
#             'Other_Assets_%_2016','Other_Assets_%_2017','Other_Assets_%_2018','Other_Assets_%_2019','Other_Assets_%_2020',
#             'MarketCap','CapFlag','LageCap','MidCap','SmallCap','TinyCap','Industry',
#             '2020_01_CLOSE','2020_02_CLOSE','2020_03_CLOSE','2020_04_CLOSE','2020_05_CLOSE','2020_06_CLOSE','2020_07_CLOSE','2020_08_CLOSE','2020_09_CLOSE','2020_10_CLOSE','2020_11_CLOSE','2020_12_CLOSE',
#             '2021']

# df_2013 = df_merge[list_2013]
# df_2013['Year'] = '2013'
# df_2014 = df_merge[list_2014]
# df_2014['Year'] = '2014'
df_2015 = df_merge[list_2015]
df_2015['Year'] = '2015'
df_2016 = df_merge[list_2016]
df_2016['Year'] = '2016'
df_2017 = df_merge[list_2017]
df_2017['Year'] = '2017'
df_2018 = df_merge[list_2018]
df_2018['Year'] = '2018'
df_2019 = df_merge[list_2019]
df_2019['Year'] = '2019'
# df_2020 = df_merge[list_2020]
# df_2020['Year'] = '2020'
# df_2021 = df_merge[list_2021]
# df_2021['Year'] = '2021'

col_names = ['Tikr','Sales_yr_1','Sales_yr_2','Sales_yr_3','Sales_yr_4','Sales_yr_5',
            'Sales_yr_2_YoY','Sales_yr_3_YoY','Sales_yr_4_YoY','Sales_yr_5_YoY',
            'Revenue_yr_1','Revenue_yr_2','Revenue_yr_3','Revenue_yr_4','Revenue_yr_5',
            'Revenue_yr_2_YoY','Revenue_yr_3_YoY','Revenue_yr_4_YoY','Revenue_yr_5_YoY',
            'Expenses_yr_1','Expenses_yr_2','Expenses_yr_3','Expenses_yr_4','Expenses_yr_5',
            'Expenses_yr_2_YoY','Expenses_yr_3_YoY','Expenses_yr_4_YoY','Expenses_yr_5_YoY',
            'Net_Profit_yr_1','Net_Profit_yr_2','Net_Profit_yr_3','Net_Profit_yr_4','Net_Profit_yr_5',
            'Net_Profit_yr_2_YoY','Net_Profit_yr_3_YoY','Net_Profit_yr_4_YoY','Net_Profit_yr_5_YoY',
            'Operating_Profit_yr_1','Operating_Profit_yr_2','Operating_Profit_yr_3','Operating_Profit_yr_4','Operating_Profit_yr_5',
            'Operating_Profit_yr_2_YoY','Operating_Profit_yr_3_YoY','Operating_Profit_yr_4_YoY','Operating_Profit_yr_5_YoY',
            'EPS_in_Rs_yr_1','EPS_in_Rs_yr_2','EPS_in_Rs_yr_3','EPS_in_Rs_yr_4','EPS_in_Rs_yr_5',
            'EPS_in_Rs_yr_2_YoY','EPS_in_Rs_yr_3_YoY','EPS_in_Rs_yr_4_YoY','EPS_in_Rs_yr_5_YoY',
            'Tax__yr_1','Tax__yr_2','Tax__yr_3','Tax__yr_4','Tax__yr_5',
            'Tax__yr_2_YoY','Tax__yr_3_YoY','Tax__yr_4_YoY','Tax__yr_5_YoY',
            'OPM__yr_1','OPM__yr_2','OPM__yr_3','OPM__yr_4','OPM__yr_5',
            'OPM__yr_2_YoY','OPM__yr_3_YoY','OPM__yr_4_YoY','OPM__yr_5_YoY',
            'Other_Assets_yr_1','Other_Assets_yr_2','Other_Assets_yr_3','Other_Assets_yr_4','Other_Assets_yr_5',
            'Other_Assets_yr_2_YoY','Other_Assets_yr_3_YoY','Other_Assets_yr_4_YoY','Other_Assets_yr_5_YoY',
            'Total_Liabilities_yr_1','Total_Liabilities_yr_2','Total_Liabilities_yr_3','Total_Liabilities_yr_4','Total_Liabilities_yr_5',
            'Total_Liabilities_yr_2_YoY','Total_Liabilities_yr_3_YoY','Total_Liabilities_yr_4_YoY','Total_Liabilities_yr_5_YoY',
            'Share_Capital_yr_1','Share_Capital_yr_2','Share_Capital_yr_3','Share_Capital_yr_4','Share_Capital_yr_5',
            'Share_Capital_yr_2_YoY','Share_Capital_yr_3_YoY','Share_Capital_yr_4_YoY','Share_Capital_yr_5_YoY',
            'Investments_yr_1','Investments_yr_2','Investments_yr_3','Investments_yr_4','Investments_yr_5',
            'Investments_yr_2_YoY','Investments_yr_3_YoY','Investments_yr_4_YoY','Investments_yr_5_YoY',
            'Borrowings_yr_1','Borrowings_yr_2','Borrowings_yr_3','Borrowings_yr_4','Borrowings_yr_5',
            'Borrowings_yr_2_YoY','Borrowings_yr_3_YoY','Borrowings_yr_4_YoY','Borrowings_yr_5_YoY',
            'Reserves_yr_1','Reserves_yr_2','Reserves_yr_3','Reserves_yr_4','Reserves_yr_5',
            'Reserves_yr_2_YoY','Reserves_yr_3_YoY','Reserves_yr_4_YoY','Reserves_yr_5_YoY',
            'Total_Assets_yr_1','Total_Assets_yr_2','Total_Assets_yr_3','Total_Assets_yr_4','Total_Assets_yr_5',
            'Total_Assets_yr_2_YoY','Total_Assets_yr_3_YoY','Total_Assets_yr_4_YoY','Total_Assets_yr_5_YoY',
            'Fixed_Assets_yr_1','Fixed_Assets_yr_2','Fixed_Assets_yr_3','Fixed_Assets_yr_4','Fixed_Assets_yr_5',
            'Fixed_Assets_yr_2_YoY','Fixed_Assets_yr_3_YoY','Fixed_Assets_yr_4_YoY','Fixed_Assets_yr_5_YoY',
            'Other_Liabilities_yr_1','Other_Liabilities_yr_2','Other_Liabilities_yr_3','Other_Liabilities_yr_4','Other_Liabilities_yr_5',
            'Other_Liabilities_yr_2_YoY','Other_Liabilities_yr_3_YoY','Other_Liabilities_yr_4_YoY','Other_Liabilities_yr_5_YoY',
            'CWIP_yr_1','CWIP_yr_2','CWIP_yr_3','CWIP_yr_4','CWIP_yr_5',
            'CWIP_yr_2_YoY','CWIP_yr_3_YoY','CWIP_yr_4_YoY','CWIP_yr_5_YoY',
            'Cash_from_Investing_Activity_yr_1','Cash_from_Investing_Activity_yr_2','Cash_from_Investing_Activity_yr_3','Cash_from_Investing_Activity_yr_4','Cash_from_Investing_Activity_yr_5',
            'Cash_from_Investing_Activity_yr_2_YoY','Cash_from_Investing_Activity_yr_3_YoY','Cash_from_Investing_Activity_yr_4_YoY','Cash_from_Investing_Activity_yr_5_YoY',
            'Cash_from_Operating_Activity_yr_1','Cash_from_Operating_Activity_yr_2','Cash_from_Operating_Activity_yr_3','Cash_from_Operating_Activity_yr_4','Cash_from_Operating_Activity_yr_5',
            'Cash_from_Operating_Activity_yr_2_YoY','Cash_from_Operating_Activity_yr_3_YoY','Cash_from_Operating_Activity_yr_4_YoY','Cash_from_Operating_Activity_yr_5_YoY',
            'Cash_from_Financing_Activity_yr_1','Cash_from_Financing_Activity_yr_2','Cash_from_Financing_Activity_yr_3','Cash_from_Financing_Activity_yr_4','Cash_from_Financing_Activity_yr_5',
            'Cash_from_Financing_Activity_yr_2_YoY','Cash_from_Financing_Activity_yr_3_YoY','Cash_from_Financing_Activity_yr_4_YoY','Cash_from_Financing_Activity_yr_5_YoY',
            'Net_Cash_Flow_yr_1','Net_Cash_Flow_yr_2','Net_Cash_Flow_yr_3','Net_Cash_Flow_yr_4','Net_Cash_Flow_yr_5',
            'Net_Cash_Flow_yr_2_YoY','Net_Cash_Flow_yr_3_YoY','Net_Cash_Flow_yr_4_YoY','Net_Cash_Flow_yr_5_YoY',
            'Share_Capital_%_yr_1','Share_Capital_%_yr_2','Share_Capital_%_yr_3','Share_Capital_%_yr_4','Share_Capital_%_yr_5',
            'Reserves_%_yr_1','Reserves_%_yr_2','Reserves_%_yr_3','Reserves_%_yr_4','Reserves_%_yr_5',
            'Borrowings_%_yr_1','Borrowings_%_yr_2','Borrowings_%_yr_3','Borrowings_%_yr_4','Borrowings_%_yr_5',
            'Other_Liabilities_%_yr_1','Other_Liabilities_%_yr_2','Other_Liabilities_%_yr_3','Other_Liabilities_%_yr_4','Other_Liabilities_%_yr_5',
            'Fixed_Assets_%_yr_1','Fixed_Assets_%_yr_2','Fixed_Assets_%_yr_3','Fixed_Assets_%_yr_4','Fixed_Assets_%_yr_5',
            'CWIP_%_yr_1','CWIP_%_yr_2','CWIP_%_yr_3','CWIP_%_yr_4','CWIP_%_yr_5',
            'Investments_%_yr_1','Investments_%_yr_2','Investments_%_yr_3','Investments_%_yr_4','Investments_%_yr_5',
            'Other_Assets_%_yr_1','Other_Assets_%_yr_2','Other_Assets_%_yr_3','Other_Assets_%_yr_4','Other_Assets_%_yr_5',
            'MarketCap','CapFlag','LageCap','MidCap','SmallCap','TinyCap','Industry',
            'Trend_lag_01','Trend_lag_02','Trend_lag_03','Trend_lag_04','Trend_lag_05','Trend_lag_06','Trend_lag_07','Trend_lag_08','Trend_lag_09','Trend_lag_10','Trend_lag_11','Trend_lag_12',
            'Return','Year']

#df_2013.columns = col_names
#df_2014.columns = col_names
df_2015.columns = col_names
df_2016.columns = col_names
df_2017.columns = col_names
df_2018.columns = col_names
df_2019.columns = col_names
# df_2020.columns = col_names
# df_2021.columns = col_names

# df_append = pd.concat([df_2015.reset_index(drop=True), df_2016.reset_index(drop=True), df_2017.reset_index(drop=True),
#                        df_2018.reset_index(drop=True), df_2019.reset_index(drop=True), df_2020.reset_index(drop=True)], axis=1)

#df_2013,
df_append = pd.concat([df_2015, df_2016, df_2017, df_2018, df_2019], ignore_index=True) #, df_2020, df_2021
#df_2020, Removed 2020 from the data purposefully in order to remove the bias that gets added due to Covid

df_append.head()
#df_append.shape

# Getting rid of revenue columns.. since it is equivalent to sales for Financial Companies
df_append['Sales_yr_1'] = df_append[['Sales_yr_1','Revenue_yr_1']].sum(axis=1)
df_append['Sales_yr_2'] = df_append[['Sales_yr_2','Revenue_yr_2']].sum(axis=1)
df_append['Sales_yr_3'] = df_append[['Sales_yr_3','Revenue_yr_3']].sum(axis=1)
df_append['Sales_yr_4'] = df_append[['Sales_yr_4','Revenue_yr_3']].sum(axis=1)
df_append['Sales_yr_5'] = df_append[['Sales_yr_5','Revenue_yr_3']].sum(axis=1)

df_append['Sales_yr_2_YoY'] = df_append[['Sales_yr_2_YoY','Revenue_yr_2_YoY']].sum(axis=1)
df_append['Sales_yr_3_YoY'] = df_append[['Sales_yr_3_YoY','Revenue_yr_3_YoY']].sum(axis=1)
df_append['Sales_yr_4_YoY'] = df_append[['Sales_yr_4_YoY','Revenue_yr_4_YoY']].sum(axis=1)
df_append['Sales_yr_5_YoY'] = df_append[['Sales_yr_5_YoY','Revenue_yr_5_YoY']].sum(axis=1)

df_append.drop(['Revenue_yr_1','Revenue_yr_2','Revenue_yr_3','Revenue_yr_4','Revenue_yr_5',
                'Revenue_yr_2_YoY','Revenue_yr_3_YoY','Revenue_yr_4_YoY','Revenue_yr_5_YoY'],axis=1,inplace=True)

df_append[df_append['Tikr'] == 'RELIANCE']

# Mapping nifty returns to subtract from 
df_bhav_return[df_bhav_return['Tikr'] == 'NIFTY50']

df_nifty_returns = df_bhav_return[df_bhav_return['Tikr'] == 'NIFTY50']
df_nifty_melt = df_nifty_returns.melt(id_vars=["Tikr"], var_name="Year", value_name="Nifty_Returns")
df_nifty_melt

df_final =  pd.merge(df_append,df_nifty_melt, how='left',left_on=['Year'], right_on=['Year'])
df_final.head()

df_final['Return_adj'] = df_final['Return'] - df_final['Nifty_Returns']
df_final.head(20)

df_final.to_csv('df_train_data_with_price.csv')

df_final[df_final['Tikr_x'] == '20MICRONS']
